refactor(examples): split chat handlers and add todo find/toggle routes
- Split sse-chat chat.rs into chat/chat_page.rs, chat/chat_send.rs,
  chat/chat_events.rs following the one-handler-per-file convention
- Add GET /todos/{id} handler (find_by_id) to todo-api
- Add PATCH /todos/{id} handler (toggle completed) to todo-api
